# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact: {}
  description: 荆楚花牌(花好月圆)游戏服务端 API
  title: 荆楚花牌 API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用邮箱和密码登录，返回 JWT token
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"token\": \"string\", \"expires_at\": \"string\", \"user\":
            {} }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "401":
          description: '{ \"error\": \"invalid credentials\" }'
          schema:
            type: object
        "403":
          description: '{ \"error\": \"email not verified\" }'
          schema:
            type: object
      summary: 用户登录
      tags:
      - 认证
  /auth/me:
    get:
      description: 返回当前已认证用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            type: object
        "404":
          description: '{ \"error\": \"user not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户并发送验证邮件
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"user_id\": \"string\", \"message\": \"验证邮件已发送，请查收邮箱\"
            }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "409":
          description: '{ \"error\": \"email already taken\" }'
          schema:
            type: object
      summary: 用户注册
      tags:
      - 认证
  /auth/resend:
    post:
      consumes:
      - application/json
      description: 重新发送邮箱验证邮件（有频率限制）
      parameters:
      - description: 邮箱
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"message\": \"如果该邮箱已注册且未验证，验证邮件已发送\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "429":
          description: '{ \"error\": \"resend throttled\" }'
          schema:
            type: object
      summary: 重发验证邮件
      tags:
      - 认证
  /auth/verify:
    get:
      description: 通过邮件中的 token 验证用户邮箱
      parameters:
      - description: 验证令牌
        in: query
        name: token
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: 验证成功页面
          schema:
            type: string
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "410":
          description: '{ \"error\": \"token already used\" }'
          schema:
            type: object
      summary: 验证邮箱
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"